Files that were modified/created in the project by us
XML files are custom layouts for the android app. Main_Activity is the form with some views and buttons, ResultActivity is the result screen. MainActivity.java is the main file. It probably wont run, missing a lot of files and configs. The .apk attached works on Pixel 5 Tiramisu in the android studio build. I think some phones wont like it in case you wanted to try; my pixel7 wont run it.